+++ src/main/java/lionking/common/LKItemAxeFire.java
@@ -1,34 +1,12 @@
 package lionking.common;
 
 import net.minecraft.block.*;
-import net.minecraft.block.material.*;
-import net.minecraft.creativetab.*;
-import net.minecraft.enchantment.*;
-import net.minecraft.entity.*;
-import net.minecraft.entity.ai.*;
-import net.minecraft.entity.item.*;
-import net.minecraft.entity.monster.*;
-import net.minecraft.entity.passive.*;
 import net.minecraft.entity.player.*;
-import net.minecraft.entity.projectile.*;
-import net.minecraft.inventory.*;
 import net.minecraft.item.*;
 import net.minecraft.item.crafting.*;
-import net.minecraft.nbt.*;
-import net.minecraft.network.packet.*;
-import net.minecraft.pathfinding.*;
-import net.minecraft.potion.*;
-import net.minecraft.server.*;
-import net.minecraft.server.management.*;
 
 import net.minecraft.stats.*;
-import net.minecraft.tileentity.*;
-import net.minecraft.util.*;
 import net.minecraft.world.*;
-import net.minecraft.world.biome.*;
-import net.minecraft.world.gen.feature.*;
-import net.minecraft.world.gen.layer.*;
-import net.minecraft.world.storage.*;
 
 import net.minecraftforge.common.*;
 
@@ -58,9 +36,9 @@ public class LKItemAxeFire extends LKItemAxe {
 					world.setBlockToAir(i, j, k);
 				}
 				for (int l = 0; l < 6; l++) {
-					double d = (double) ((float) i + itemRand.nextFloat());
-					double d1 = (double) ((float) j + itemRand.nextFloat());
-					double d2 = (double) ((float) k + itemRand.nextFloat());
+					double d = i + itemRand.nextFloat();
+					double d1 = j + itemRand.nextFloat();
+					double d2 = k + itemRand.nextFloat();
 					world.spawnParticle("flame", d, d1, d2, 0.0D, 0.0D, 0.0D);
 				}
 				entityplayer.triggerAchievement(LKAchievementList.fireTool);
